requirements.txt
streamlit>=1.28.0
langgraph>=0.0.40
langchain-google-genai>=1.0.0
yfinance>=0.2.18
pandas>=1.5.0
requests>=2.28.0
sec-edgar-api>=1.1.0
alpha-vantage>=2.3.1
python-dotenv>=1.0.0
plotly>=5.15.0
beautifulsoup4>=4.12.0